+++ doc/RELEASE_NOTES.md
@@ -14,3 +14,4 @@ v1.1.0
 ### New features
 
 - Added support for `plantuml` diagrams in Doxygen markup.
+